requirements.txt
phidata
openai
python-environ
langchainhub==0.1.15 
langchain_experimental
langchain_openai
faiss-cpu
